vis tests passing
