+++ tests/utils/conftest.py
@@ -11,10 +11,10 @@ import os
 import pytest
 import numpy as np
 
-from gnss_lib_py.utils.time_conversions import gps_millis_to_datetime
 from gnss_lib_py.navdata.navdata import NavData
 from gnss_lib_py.parsers.google_decimeter import AndroidDerived2022, AndroidGroundTruth2022
 from gnss_lib_py.parsers.rinex_nav import get_time_cropped_rinex
+from gnss_lib_py.navdata.operations import loop_time
 
 def pytest_collection_modifyitems(items):
     """Run ephemeris download tests after all other tests.
@@ -221,7 +221,7 @@ def fixture_derived_gps_l1_reversed(android_gps_l1):
         instance.
     """
     android_gps_l1_reversed = NavData()
-    for _, _, measure_frame in android_gps_l1.loop_time('gps_millis', delta_t_decimals=-2):
+    for _, _, measure_frame in loop_time(android_gps_l1,'gps_millis', delta_t_decimals=-2):
         if len(android_gps_l1_reversed)==0:
             android_gps_l1_reversed = measure_frame
         else:
@@ -348,7 +348,7 @@ def fixture_gps_measurement_frames(all_gps_ephem, android_gps_l1):
         received Android measurements and SV states. The lists are
         indexed by discrete time indices.
     """
-    android_frames = android_gps_l1.loop_time('gps_millis', delta_t_decimals=-2)
+    android_frames = loop_time(android_gps_l1,'gps_millis', delta_t_decimals=-2)
     ephems = []
     frames = []
     sv_states = []
