+++ tests/visualizations/test_plot_map.py
@@ -5,6 +5,11 @@
 __authors__ = "D. Knowles"
 __date__ = "22 Jun 2022"
 
+import pytest
+
+import plotly.graph_objects as go
+from gnss_lib_py.visualizations import plot_map
+
 def test_plot_map(gtruth, state_estimate):
     """Test for plotting map.
 
@@ -20,22 +25,22 @@ def test_plot_map(gtruth, state_estimate):
 
     """
 
-    fig = viz.plot_map(gtruth, state_estimate, save=False)
+    fig = plot_map.plot_map(gtruth, state_estimate, save=False)
     assert isinstance(fig, go.Figure)
 
-    figs = viz.plot_map(gtruth, state_estimate,sections=3, save=False)
+    figs = plot_map.plot_map(gtruth, state_estimate,sections=3, save=False)
     for fig in figs:
         assert isinstance(fig, go.Figure)
 
     with pytest.raises(TypeError) as excinfo:
-        viz.plot_map([], state_estimate, save=False)
+        plot_map.plot_map([], state_estimate, save=False)
     assert "NavData" in str(excinfo.value)
     assert "Input" in str(excinfo.value)
 
     for row in ["lat_rx_wls_deg","lon_rx_wls_deg"]:
         state_removed = state_estimate.remove(rows=row)
         with pytest.raises(KeyError) as excinfo:
-            viz.plot_map(gtruth, state_removed, save=False)
+            plot_map.plot_map(gtruth, state_removed, save=False)
         assert row.replace("rx_wls","*") in str(excinfo.value)
         assert "Missing" in str(excinfo.value)
 
@@ -43,6 +48,6 @@ def test_plot_map(gtruth, state_estimate):
         state_double = state_estimate.copy()
         state_double[row.replace("rx","2")] = state_double[row]
         with pytest.raises(KeyError) as excinfo:
-            viz.plot_map(gtruth, state_double, save=False)
+            plot_map.plot_map(gtruth, state_double, save=False)
         assert row.replace("rx_wls","*") in str(excinfo.value)
         assert "More than 1" in str(excinfo.value)
