+++ tests/visualizations/test_plot_skyplot.py
@@ -5,8 +5,19 @@
 __authors__ = "D. Knowles"
 __date__ = "22 Jun 2022"
 
+import os
+
 import pytest
+import numpy as np
+import matplotlib as mpl
 from pytest_lazyfixture import lazy_fixture
+
+from gnss_lib_py.navdata.navdata import NavData
+from gnss_lib_py.navdata.operations import find_wildcard_indexes
+from gnss_lib_py.visualizations import style
+from gnss_lib_py.visualizations import plot_skyplot
+from gnss_lib_py.utils.coordinates import geodetic_to_ecef
+from gnss_lib_py.parsers.google_decimeter import AndroidDerived2022
 
 @pytest.mark.parametrize('navdata',[
                                     # lazy_fixture('derived_2022'),
@@ -36,37 +47,37 @@ def test_plot_skyplot(navdata, state_estimate):
             navdata["x_sv_m",col_idx] = np.nan
 
     # don't save figures
-    fig = viz.plot_skyplot(navdata.copy(), state_estimate, save=False)
-    viz.close_figures(fig)
+    fig = plot_skyplot.plot_skyplot(navdata.copy(), state_estimate, save=False)
+    style.close_figures(fig)
 
     with pytest.raises(TypeError) as excinfo:
-        viz.plot_skyplot(navdata.copy(), state_estimate, save=True, prefix=1)
+        plot_skyplot.plot_skyplot(navdata.copy(), state_estimate, save=True, prefix=1)
     assert "Prefix" in str(excinfo.value)
 
     with pytest.raises(TypeError) as excinfo:
-        viz.plot_skyplot("derived", "raw_pr_m", save=True)
+        plot_skyplot.plot_skyplot("derived", "raw_pr_m", save=True)
     assert "NavData" in str(excinfo.value)
 
     for row in ["x_sv_m","y_sv_m","z_sv_m","gps_millis"]:
         derived_removed = navdata.remove(rows=row)
         with pytest.raises(KeyError) as excinfo:
-            viz.plot_skyplot(derived_removed, state_estimate, save=False)
+            plot_skyplot.plot_skyplot(derived_removed, state_estimate, save=False)
         assert row in str(excinfo.value)
 
     for row in ["x_rx_m","y_rx_m","z_rx_m"]:
-        row_idx = state_estimate.find_wildcard_indexes(row[:4]+'*'+row[4:])[row[:4]+'*'+row[4:]][0]
+        row_idx = find_wildcard_indexes(state_estimate,row[:4]+'*'+row[4:])[row[:4]+'*'+row[4:]][0]
         state_removed = state_estimate.remove(rows=row_idx)
         with pytest.raises(KeyError) as excinfo:
-            viz.plot_skyplot(navdata, state_removed, save=False)
+            plot_skyplot.plot_skyplot(navdata, state_removed, save=False)
         assert row[:4]+'*'+row[4:] in str(excinfo.value)
         assert "Missing" in str(excinfo.value)
 
     for row in ["x_rx_m","y_rx_m","z_rx_m"]:
         state_double = state_estimate.copy()
-        row_idx = state_estimate.find_wildcard_indexes(row[:4]+'*'+row[4:])[row[:4]+'*'+row[4:]][0]
+        row_idx = find_wildcard_indexes(state_estimate,row[:4]+'*'+row[4:])[row[:4]+'*'+row[4:]][0]
         state_double[row_idx.replace("rx_","rx_gt_")] = state_double[row_idx]
         with pytest.raises(KeyError) as excinfo:
-            viz.plot_skyplot(navdata, state_double, save=False)
+            plot_skyplot.plot_skyplot(navdata, state_double, save=False)
         assert row[:4]+'*'+row[4:] in str(excinfo.value)
         assert "More than 1" in str(excinfo.value)
 
@@ -91,7 +102,7 @@ def test_skyplot_trim(root_path):
     receiver_state["y_rx_m"] = y_rx_m
     receiver_state["z_rx_m"] = z_rx_m
 
-    fig = viz.plot_skyplot(sp3,receiver_state)
+    fig = plot_skyplot.plot_skyplot(sp3,receiver_state)
     # verify that two line segments were removed. Should be 57 not 59
     # after trimming the two separated ones.
     for child in fig.get_children():
@@ -99,9 +110,9 @@ def test_skyplot_trim(root_path):
             for grandchild in child.get_children():
                 if isinstance(grandchild,mpl.collections.LineCollection):
                     assert len(grandchild.get_array()) == 57
-    viz.close_figures()
+    style.close_figures()
 
-    fig = viz.plot_skyplot(sp3,receiver_state,trim_options={"az" : 95.})
+    fig = plot_skyplot.plot_skyplot(sp3,receiver_state,trim_options={"az" : 95.})
     # verify that only one line segment was removed. Should be 58 not 59
     # after trimming the one larger than 95 degrees in azimuth separated ones.
     for child in fig.get_children():
@@ -109,9 +120,9 @@ def test_skyplot_trim(root_path):
             for grandchild in child.get_children():
                 if isinstance(grandchild,mpl.collections.LineCollection):
                     assert len(grandchild.get_array()) == 58
-    viz.close_figures()
+    style.close_figures()
 
-    fig = viz.plot_skyplot(sp3,receiver_state,trim_options={"gps_millis" : 3.5E6})
+    fig = plot_skyplot.plot_skyplot(sp3,receiver_state,trim_options={"gps_millis" : 3.5E6})
     # verify that only one line segment was removed. Should be 58 not 59
     # after trimming the one larger than 95 degrees in azimuth separated ones.
     for child in fig.get_children():
@@ -119,9 +130,9 @@ def test_skyplot_trim(root_path):
             for grandchild in child.get_children():
                 if isinstance(grandchild,mpl.collections.LineCollection):
                     assert len(grandchild.get_array()) == 57
-    viz.close_figures()
+    style.close_figures()
 
 
     with pytest.raises(TypeError) as excinfo:
-        viz.plot_skyplot(sp3, receiver_state, step=20.1)
+        plot_skyplot.plot_skyplot(sp3, receiver_state, step=20.1)
     assert "step" in str(excinfo.value)
